added module to create product data object
Notify-me work has created more variants (based on changing size,
changing stock, sales, etc). To deal with these variants, we're
creating a javascript object that accesses the data at once, and can be
passed around. We don't need to check if a product is "on sale" in
multiple modules...
I've not implemented on existing modules yet. This can be done anytime
moving forward, and is recommended to improve performance.