Created toggleHidden module to handle field hiding
Helper module to handle hiding and un-hiding of form fields.
This was caused by the browser's native form validation checking hidden fields, so as a workaround we apply the "disabled" property to fields when hidden, and remove the propery when un-hidden. The helper was created to provide a central place for these changes.

This will be used on the notify-me page immediatley, with some changes in validator to ignore hidden fields.